Refactoring into baseCommand
