+++ connectors/connector-postgresql/connector_postgresql/commands/createTable.py
@@ -1,7 +1,8 @@
 import json
-import psycopg2
 
-class CreateTable:
+from connector_postgresql.baseCommand import BaseCommand
+
+class CreateTable(BaseCommand):
     """CreateTable."""
 
     def __init__(self, table_name: str, schema: str):
@@ -16,19 +17,9 @@ class CreateTable:
         # https://www.psycopg.org/docs/usage.html#passing-parameters-to-sql-queries
         sql = f"CREATE TABLE {self.table_name} ({columns});"
 
-        with psycopg2.connect(self._get_db_connection_str(config)) as conn:
-            with conn.cursor() as cursor:
-                cursor.execute(sql)
-                conn.commit
+        (response, status) = self.perform_query(sql, config)
 
-        return {"response": f'{{"result": "created table {self.table_name}"}}', "status": 200, "mimetype": "application/json"}
-
-    def _get_db_connection_str(self, config):
-        database = config["POSTGRESQL_DB_NAME"]
-        username = config["POSTGRESQL_USER_NAME"]
-        password = config["POSTGRESQL_PASSWORD"]
-        
-        return f"dbname={database} user={username} password={password}"
+        return {"response": response, "status": status, "mimetype": "application/json"}
 
     def _column_definitions(self, schema):
         def column_defintion(column):
